Split notes into another module
